requirements.txt
streamlit
langchain
cassio
astrapy
together
wikipedia-api
google-api-python-client
python-dotenv
graphviz
plotly